pyproject.toml
[tool.poetry]
name = "branch-splitter"
version = "0.1.0"
description = ""
authors = ["Jordan Aspinall <jordan.aspinall@pm.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.12.5"
gitpython = "^3.1.43"
pytest = "^8.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
